support dynamically sized receive buffers
This is the core of HPN-SSH. Please note that this patch will not
necessarily improve all connections but only those that were being hampered
by undersized buffers. While  optimal improvement requires this patch on
both sides of the connection many users see enhancement as long as the
patch is applied on the side of the connection receiving bulk data.